Implement passport support

+++ controllers/TasksController.js
@@ -3,13 +3,14 @@ module.exports = (app) => {
     let TasksController = {
         index: (req, res) => {
             // "/tasks": Lista tarefas
-            Tasks.findAll({})
+            Tasks.findAll({where: {user_id: req.user.id}})
                 .then((result) => res.json(result))
                 .catch((error) => res.status(412).json({msg: error.message}));
         },
 
         create: (req, res) => {
-            // "/tasks": Cadastra nova tarefa
+            // "/tasks": Cadastra nova tarefa\
+            req.body.user_id = req.user.id;
             Tasks.create(req.body)
                 .then((result) => res.json(result))
                 .catch((error) => res.status(412).json({msg: error.message}));
@@ -17,7 +18,7 @@ module.exports = (app) => {
 
         findOne: (req, res) => {
             // "/tasks/1": Consulta uma tarefa pelo id
-            Tasks.findOne({where: req.params})
+            Tasks.findOne({where: {id: req.params.id, user_id: req.user.id}})
                 .then((result) => {
                     if (result) {
                         res.json(result);
@@ -30,7 +31,7 @@ module.exports = (app) => {
 
         updateOne: (req, res) => {
             // "/tasks/1": Atualiza a tarefa pelo id
-            Tasks.update(req.body, {where: req.params})
+            Tasks.update(req.body, {where: {id: req.params.id, user_id: req.user.id}})
                 .then(result => res.sendStatus(204))
                 .catch(error => res.status(412).json({msg: error.message}));
         
@@ -38,7 +39,7 @@ module.exports = (app) => {
 
         delete: (req, res) => {
         // "/tasks/1": Deleta a tarefa pelo id
-        Tasks.destroy({where: req.params})
+        Tasks.destroy({where: {id: req.params.id, user_id: req.user.id}})
             .then(result => res.sendStatus(204))
             .catch(error => res.status(412).json({msg: error.message}));
